requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
python-docx
tqdm
sqlite3
httpx
pytest
pytest-asyncio
numpy
sqlalchemy
rich
rapidfuzz
